add support for changing genl family with environment variables
Multiple platforms can not register the same family, so ghibli registers
with different ones. This enables the CLI to change the family, and
work with ghibli (on a maserati host).

+++ src/mchp_ui_qos.h
@@ -148,7 +148,7 @@ struct mchp_frer_iflow_cfg {
 	u16 ms_id;       /* Member stream base ID */
 	bool generation; /* Enable/disable sequence generation */
 	bool pop;        /* Pop R-tag */
-	u8 split_mask;   /* Egress ports to split into */
+	u16 split_mask;   /* Egress ports to split into */
 };
 
 /* Ingress flow configuration */
